fix test failures due to presence of response columns and function not found.
